[rename🚚] AOP 패키지 위치 이동에 따른 변경 내용(#43)
